LayoutTests/http/tests/media/media-stream/disconnected-frame-permission-denied-expected.txt

Tests that when a getUserMedia request is made, permission is denied and its Frame is disconnected before the promise is resolved, the promise never resolves.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.name is "NotAllowedError"

PASS No callbacks invoked
PASS successfullyParsed is true

TEST COMPLETE
